requirements.txt
Flask==0.10.1
Jinja2==2.7.2
PyYAML==3.11
Werkzeug==0.9.4
mock==1.0.1
nose==1.3.0
rednose==0.4.1
gunicorn==19.1.0
SQLAlchemy=0.9.7
Flask-SQLAlchemy=2.0